+++ library/src/main/java/de/mrapp/android/tabswitcher/layout/AbstractTabRecyclerAdapter.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 - 2017 Michael Rapp
+ * Copyright 2016 - 2018 Michael Rapp
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
@@ -502,9 +502,10 @@ public abstract class AbstractTabRecyclerAdapter
     @Override
     public final void onTabAdded(final int index, @NonNull final Tab tab,
                                  final int previousSelectedTabIndex, final int selectedTabIndex,
+                                 final boolean selectionChanged,
                                  final boolean switcherVisibilityChanged,
                                  @NonNull final Animation animation) {
-        if (previousSelectedTabIndex != selectedTabIndex) {
+        if (selectionChanged) {
             adaptAllSelectionStates();
         }
     }
@@ -512,8 +513,9 @@ public abstract class AbstractTabRecyclerAdapter
     @Override
     public final void onAllTabsAdded(final int index, @NonNull final Tab[] tabs,
                                      final int previousSelectedTabIndex, final int selectedTabIndex,
+                                     final boolean selectionChanged,
                                      @NonNull final Animation animation) {
-        if (previousSelectedTabIndex != selectedTabIndex) {
+        if (selectionChanged) {
             adaptAllSelectionStates();
         }
     }
@@ -521,8 +523,9 @@ public abstract class AbstractTabRecyclerAdapter
     @Override
     public final void onTabRemoved(final int index, @NonNull final Tab tab,
                                    final int previousSelectedTabIndex, final int selectedTabIndex,
+                                   final boolean selectionChanged,
                                    @NonNull final Animation animation) {
-        if (previousSelectedTabIndex != selectedTabIndex) {
+        if (selectionChanged) {
             adaptAllSelectionStates();
         }
     }
@@ -535,6 +538,11 @@ public abstract class AbstractTabRecyclerAdapter
 
     @Override
     public void onPaddingChanged(final int left, final int top, final int right, final int bottom) {
+
+    }
+
+    @Override
+    public void onApplyPaddingToTabsChanged(final boolean applyPaddingToTabs) {
 
     }