+++ library/src/main/java/de/mrapp/android/tabswitcher/layout/tablet/TabletContentRecyclerAdapterWrapper.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2016 - 2017 Michael Rapp
+ * Copyright 2016 - 2018 Michael Rapp
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
  * in compliance with the License. You may obtain a copy of the License at
@@ -248,6 +248,7 @@ public class TabletContentRecyclerAdapterWrapper extends AbstractViewRecycler.Ad
     @Override
     public final void onTabAdded(final int index, @NonNull final Tab tab,
                                  final int previousSelectedTabIndex, final int selectedTabIndex,
+                                 final boolean selectionChanged,
                                  final boolean switcherVisibilityChanged,
                                  @NonNull final Animation animation) {
 
@@ -256,6 +257,7 @@ public class TabletContentRecyclerAdapterWrapper extends AbstractViewRecycler.Ad
     @Override
     public final void onAllTabsAdded(final int index, @NonNull final Tab[] tabs,
                                      final int previousSelectedTabIndex, final int selectedTabIndex,
+                                     final boolean selectionChanged,
                                      @NonNull final Animation animation) {
 
     }
@@ -263,6 +265,7 @@ public class TabletContentRecyclerAdapterWrapper extends AbstractViewRecycler.Ad
     @Override
     public final void onTabRemoved(final int index, @NonNull final Tab tab,
                                    final int previousSelectedTabIndex, final int selectedTabIndex,
+                                   final boolean selectionChanged,
                                    @NonNull final Animation animation) {
 
     }
@@ -276,6 +279,11 @@ public class TabletContentRecyclerAdapterWrapper extends AbstractViewRecycler.Ad
     @Override
     public final void onPaddingChanged(final int left, final int top, final int right,
                                        final int bottom) {
+
+    }
+
+    @Override
+    public final void onApplyPaddingToTabsChanged(final boolean applyPaddingToTabs) {
 
     }